requirements.txt
attrs==19.3.0
awsebcli==3.18.2
bcrypt==3.1.7
blessed==1.17.9
blinker==1.4
boto3==1.14.43
botocore==1.17.43
cached-property==1.5.1
cachetools==4.1.1
cement==2.8.2
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
dateutils==0.6.8
decorator==4.4.2
dnspython==2.0.0
docker==4.2.2
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JSON==0.3.4
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
future==0.16.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.7
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mongoengine==0.20.0
mongoengine-goodjson==1.1.8
oauth2client==4.1.3
oauthlib==3.1.0
paramiko==2.7.1
passlib==1.7.2
pathspec==0.5.9
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dateutil==2.8.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
